Registration API update
1. Add new API method POST /v3/registrations/me/{eventId} which is just an alias for the existing POST /v3/registrations, that can be used to for self-registration. It doesn't take any parameters, so to customize registration the existing PUT /v3/registrations/{id} must be used.

2. Some refactoring made, two controllers merged into one: RegistrationReadController and RegistrationUpdateController merged into RegistrationsController.
